initial commit
This is the version for "generic" esp32. tested on an ESP32-WROOM (Joy-IT SBC-NodeMCU-ESP32)

It makes use of the deep sleep function. it defines 4 GPIO Pins which wake it up and execute commands, then goes to sleep again..
by default I have used GPIO12, 13, 14 and 27. It needs to be an RTC_GPIO. More info about deep sleep and wakeup here: https://randomnerdtutorials.com/esp32-deep-sleep-arduino-ide-wake-up-sources/

To change GPIO Pins, you need to change #define WakeupPins and the pins in the function "sendCommand"